requirements.txt
openai==1.30.5
uvicorn==0.30.1
fastapi==0.111.0
pydantic==2.7.2
PyYAML~=6.0.1
httpx==0.27.0
jsonref~=1.1.0
starlette==0.37.2
sse-starlette==2.1.0
python-dotenv==1.0.1
pytest==8.2.1
pydantic-settings==2.2.1
sentry-sdk[fastapi]==2.3.1
pre-commit==3.7.1
